+++ CHANGELOG.md
@@ -11,6 +11,10 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [0.0.11] - 2025-05-25
 
+### Changed
+
+- Use Jinja to render the script.
+
 ### Fixed
 
 - Fix label generation when a prefix is specified.
@@ -100,4 +104,4 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 First version.
 
-[unreleased]: https://github.com/Tatsh/gendisc/-/compare/v0.0.10...master
+[unreleased]: https://github.com/Tatsh/gendisc/-/compare/v0.0.11...master